Introduces CRCs (#98)
* Introduces CRCs

A `crc` feature has been added that brings in the `crc` crate and a new `CrcModifier` flavour. This modifier flavour is similar to the COBS implementation in how it wraps another flavour.

* Convenience functions

Convenience functions for serde usage

* Optionally include paste

The paste crate is included now only if `use-crc` feature is enabled. This may help improve build performance where CRCs are not required.

* Fix use of paste macro

This uses the correct digest sizes, and adds a test for
a random CRC8 variant.

---------

Co-authored-by: James Munns <james@onevariable.com>